requirements.txt
mysql-connector-python
pandas
matplotlib
torch
torchvision
requests
Flask